Move fakes to be in core package
