+++ core/fake/log_verifier.go
